correcttool/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.2
distlib==0.3.4
distro==1.9.0
et-xmlfile==1.1.0
filelock==3.4.2
frozenlist==1.4.1
h11==0.13.0
httpcore==1.0.5
httpx==0.27.0
idna==3.3
jaconv==0.3.4
mecab-python3==1.0.9
multidict==6.0.5
numpy==1.26.4
openai==1.37.1
openpyxl==3.1.2
pandas==2.2.2
platformdirs==2.4.1
pydantic==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.5
starlette==0.17.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.13.1
yarl==1.9.4
